Edit help files, make >tag non-public.

+++ avebot.py
@@ -235,7 +235,6 @@ async def get_images(contx):
             im.save(new_name, "JPEG")
             filename = new_name
         images_to_process.append(filename)
-    avelog("{} images found to process1, repr: {}".format(len(images_to_process), repr(images_to_process)))
     return images_to_process
 
 
@@ -243,7 +242,6 @@ async def get_images(contx):
 async def sbahjify(contx):
     """Makes images hella and sweet."""
     images_to_process = await get_images(contx)
-    avelog("{} images found to process2, repr: {}".format(len(images_to_process), repr(images_to_process)))
     msg_to_send = '{}: Processing image(s).' if len(
         images_to_process) != 0 else '{}: No images found. Try linking them or uploading them directly through discord.'
     tmp = await bot.send_message(contx.message.channel, msg_to_send.format(contx.message.author.mention))
@@ -275,22 +273,22 @@ async def sbahjify(contx):
 @bot.command(pass_context=True)
 async def tag(contx):
     """Tags images. Based on tagbox."""
-    images_to_process = await get_images(contx)
-    avelog("{} images found to process3, repr: {}".format(len(images_to_process), repr(images_to_process)))
-    msg_to_send = '{}: Processing image(s).' if len(
-        images_to_process) != 0 else '{}: No images found. Try linking them or uploading them directly through discord.'
-    tmp = await bot.send_message(contx.message.channel, msg_to_send.format(contx.message.author.mention))
-    for imgtp in images_to_process:
-        avelog("Processing {}".format(imgtp))
-        post = requests.post("http://52.168.149.3:8080/tagbox/check", files={'file': open(imgtp, 'rb')})
-        postj = post.json()
-        if postj.success:
-            text = "Image successfully viewed~ [based on tagbox] \nTags:"
-            for t in postj.tags:
-                text.append("**{}** ({} confidence)\n".format(t.tag, t.confidence))
-            await bot.send_message(contx.message.channel, "{}: {}".format(contx.message.author.mention, text))
-    await asyncio.sleep(5)
-    await bot.delete_message(tmp)
+    if check_level(contx.message.author.id) in ["2", "8", "9"]:
+        images_to_process = await get_images(contx)
+        msg_to_send = '{}: Processing image(s). (this might take some time)' if len(
+            images_to_process) != 0 else '{}: No images found. Try linking them or uploading them directly through discord.'
+        tmp = await bot.send_message(contx.message.channel, msg_to_send.format(contx.message.author.mention))
+        for imgtp in images_to_process:
+            avelog("Processing {}".format(imgtp))
+            post = requests.post("http://52.168.149.3:8080/tagbox/check", files={'file': open(imgtp, 'rb')})
+            postj = post.json()
+            if postj.success:
+                text = "Image successfully viewed~ [based on tagbox] \nTags:"
+                for t in postj.tags:
+                    text.append("**{}** ({} confidence)\n".format(t.tag, t.confidence))
+                await bot.send_message(contx.message.channel, "{}: {}".format(contx.message.author.mention, text))
+        await asyncio.sleep(5)
+        await bot.delete_message(tmp)
 
 
 @bot.command()
